ghetto refresh to avoid refresh spam
