+++ src/controllers/v2/client.js
@@ -1,15 +1,7 @@
-const axios = require("axios");
-const { HIE } = require("../../../config");
 const { supportedIDTypes, areSimilar } = require("../../utils/client");
 const { getDoc, genRequestConfig } = require("../../utils/echis");
-const { reauthenticate } = require("../../middlewares/v2/auth");
-
-const axiosInstance = axios.create({ baseURL: HIE.url, timeout: 10000 });
-axiosInstance.interceptors.response.use(
-  (response) => response,
-  (error) => {
-    return reauthenticate(axiosInstance, error)
-  });
+const { AuthenticatedInstance } = require("./auth");
+const authenticatedInstance = new AuthenticatedInstance();
 
 const err = (resp, err) => {
   if (err) throw err;
@@ -38,7 +30,7 @@ const parseID = (contact) => {
 const fetchClientFromRegistry = async (contact) => {
   const params = parseID(contact);
   try {
-    const resp = await axiosInstance.get("/v1/fhir/Patient", {
+    const resp = await authenticatedInstance.instance().get("/v1/fhir/Patient", {
       params: { ...params },
     });
     if (resp.data.total <= 0) {
